refactor: unify CloudProviderCard into @freed/ui shared package
Move the canonical CloudProviderCard component (and its ProviderState /
CloudProvider types) from packages/desktop into packages/ui so that both
the PWA and desktop consume the same component.

- packages/ui: add CloudProviderCard with types; export via package.json
- packages/desktop: CloudProviderCard.tsx becomes a thin re-export;
  useCloudProviders imports ProviderState from @freed/ui instead of
  redefining it
- packages/pwa: SyncConnectDialog cloud tab now uses CloudProviderCard
  instead of the bespoke inline buttons that showed "Connect Dropbox" /
  "Connect Google Drive" (mismatched with desktop's "Dropbox" label)

Fixes label divergence: both surfaces now show "Dropbox" / "Google Drive"
as the provider name with a separate "Connect" action button.
