+++ packages/pwa/src/components/SyncConnectDialog.tsx
@@ -10,6 +10,7 @@ import {
 } from "../lib/sync";
 import { BottomSheet } from "@freed/ui/components/BottomSheet";
 import { addDebugEvent } from "@freed/ui/lib/debug-store";
+import { CloudProviderCard } from "@freed/ui/components/CloudProviderCard";
 
 // ─── OAuth PKCE helpers ───────────────────────────────────────────────────────
 
@@ -551,41 +552,14 @@ export function SyncConnectDialog({ open, onClose, initialMode = "cloud" }: Sync
 
       {mode === "cloud" && (
         <div className="mb-4 flex flex-col gap-3">
-          <button
-            onClick={() => handleCloudConnect("dropbox")}
-            disabled={cloudConnecting}
-            className="flex items-center gap-3 w-full px-4 py-3.5 bg-white/5 hover:bg-white/10 border border-[rgba(255,255,255,0.08)] hover:border-[#8b5cf6]/40 rounded-xl transition-colors disabled:opacity-50"
-          >
-            {/* Dropbox icon */}
-            <svg className="w-5 h-5 flex-shrink-0" viewBox="0 0 24 24" fill="#0061FF">
-              <path d="M6 2L0 6l6 4-6 4 6 4 6-4-6-4 6-4L6 2zM18 2l-6 4 6 4-6 4 6 4 6-4-6-4 6-4-6-4zM12 14l-6 4 6 4 6-4-6-4z"/>
-            </svg>
-            <div className="text-left">
-              <div className="text-sm font-medium text-white">Connect Dropbox</div>
-              <div className="text-xs text-[#71717a]">~2s sync latency · stored in /Apps/Freed/</div>
-            </div>
-          </button>
-
-          <button
-            onClick={() => handleCloudConnect("gdrive")}
-            disabled={cloudConnecting}
-            className="flex items-center gap-3 w-full px-4 py-3.5 bg-white/5 hover:bg-white/10 border border-[rgba(255,255,255,0.08)] hover:border-[#8b5cf6]/40 rounded-xl transition-colors disabled:opacity-50"
-          >
-            {/* Google Drive icon */}
-            <svg className="w-5 h-5 flex-shrink-0" viewBox="0 0 87.3 78" fill="none">
-              <path d="M6.6 66.85l3.85 6.65c.8 1.4 1.95 2.5 3.3 3.3L27.5 53H0c0 1.55.4 3.1 1.2 4.5z" fill="#0066DA"/>
-              <path d="M43.65 25L29.9 1.2C28.55 2 27.4 3.1 26.6 4.5L1.2 48.5A9.06 9.06 0 000 53h27.5z" fill="#00AC47"/>
-              <path d="M73.55 76.8c1.35-.8 2.5-1.9 3.3-3.3l1.6-2.75 7.65-13.25c.8-1.4 1.2-2.95 1.2-4.5H59.8l5.85 11.2z" fill="#EA4335"/>
-              <path d="M43.65 25L57.4 1.2C56.05.4 54.5 0 52.9 0H34.4c-1.6 0-3.15.45-4.5 1.2z" fill="#00832D"/>
-              <path d="M59.8 53H27.5L13.75 76.8c1.35.8 2.9 1.2 4.5 1.2h50.8c1.6 0 3.15-.4 4.5-1.2z" fill="#2684FC"/>
-              <path d="M73.4 26.5l-12.65-21.9c-.8-1.4-1.95-2.5-3.3-3.3L43.65 25 59.8 53h27.45c0-1.55-.4-3.1-1.2-4.5z" fill="#FFBA00"/>
-            </svg>
-            <div className="text-left">
-              <div className="text-sm font-medium text-white">Connect Google Drive</div>
-              <div className="text-xs text-[#71717a]">~4s sync latency · stored in app data folder</div>
-            </div>
-          </button>
-
+          {(["dropbox", "gdrive"] as const).map((provider) => (
+            <CloudProviderCard
+              key={provider}
+              provider={provider}
+              state={{ status: cloudConnecting ? "connecting" : "idle" }}
+              onConnect={handleCloudConnect}
+            />
+          ))}
           <p className="text-xs text-[#71717a] text-center pt-1">
             Your reading list is stored in your own cloud account.
             <br />
